chore: move `TestCanCreateContract` to original file
Simplifies code review
